Added TDD tests made from refactored existing tests
no sgnificant changes where needed, mostly naming conventions
and approach.

it was easer for me to refacter in a seperate file and keeping
the old tests for now might be wise if the TDD approach does
not work for whatever reason.

for now will update these tests throught further dev work.